web_app/backend/requirements.txt
fastapi
uvicorn
mangum
langgraph
google-genai
openai
boto3
pydantic
python-dotenv
firebase-admin
python-multipart
